requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-cors-headers==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
mssql-django==1.3
mysqlclient==2.2.0
numpy==1.26.0
packaging==23.1
Pillow==10.0.1
PyJWT==2.8.0
pyodbc==4.0.39
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.5
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
